Bluetooth: Host: Move BR/EDR files to host/classic
Add subfolder "subsys/bluetooth/host/classic".
Move BR/EDR source files to the subfolder.
