+++ subsys/bluetooth/host/classic/a2dp.c
@@ -23,8 +23,8 @@
 
 #include "common/assert.h"
 
-#include "hci_core.h"
-#include "conn_internal.h"
+#include "host/hci_core.h"
+#include "host/conn_internal.h"
 #include "avdtp_internal.h"
 #include "a2dp_internal.h"
