+++ subsys/bluetooth/host/classic/a2dp_internal.h
