+++ subsys/bluetooth/host/classic/avdtp.c
@@ -18,8 +18,8 @@
 #include <zephyr/bluetooth/l2cap.h>
 #include <zephyr/bluetooth/avdtp.h>
 
-#include "hci_core.h"
-#include "conn_internal.h"
+#include "host/hci_core.h"
+#include "host/conn_internal.h"
 #include "l2cap_br_internal.h"
 #include "avdtp_internal.h"
