+++ subsys/bluetooth/host/classic/avdtp_internal.h
