+++ subsys/bluetooth/host/classic/hfp_hf.c
@@ -19,8 +19,8 @@
 #include <zephyr/bluetooth/rfcomm.h>
 #include <zephyr/bluetooth/hfp_hf.h>
 
-#include "hci_core.h"
-#include "conn_internal.h"
+#include "host/hci_core.h"
+#include "host/conn_internal.h"
 #include "l2cap_br_internal.h"
 #include "rfcomm_internal.h"
 #include "at.h"
