+++ subsys/bluetooth/host/classic/hfp_internal.h
