+++ subsys/bluetooth/host/classic/rfcomm.c
@@ -22,8 +22,8 @@
 
 #include <zephyr/bluetooth/rfcomm.h>
 
-#include "hci_core.h"
-#include "conn_internal.h"
+#include "host/hci_core.h"
+#include "host/conn_internal.h"
 #include "l2cap_br_internal.h"
 #include "rfcomm_internal.h"
