+++ subsys/bluetooth/host/classic/rfcomm_internal.h
