+++ subsys/bluetooth/host/classic/sdp.c
@@ -19,8 +19,8 @@
 #include "common/bt_str.h"
 #include "common/assert.h"
 
-#include "hci_core.h"
-#include "conn_internal.h"
+#include "host/hci_core.h"
+#include "host/conn_internal.h"
 #include "l2cap_br_internal.h"
 #include "sdp_internal.h"
