+++ subsys/bluetooth/host/classic/sdp_internal.h
