+++ subsys/bluetooth/host/classic/ssp.c
@@ -15,10 +15,10 @@
 
 #include "common/bt_str.h"
 
-#include "keys.h"
+#include "host/keys.h"
 
-#include "hci_core.h"
-#include "conn_internal.h"
+#include "host/hci_core.h"
+#include "host/conn_internal.h"
 
 #define LOG_LEVEL CONFIG_BT_HCI_CORE_LOG_LEVEL
 #include <zephyr/logging/log.h>
