+++ subsys/bluetooth/host/classic/ssp.h
